==> notebooks/00_eda.ipynb <==
# Sports Betting Dataset

Current challenge is to build a recommendation engine to suggest betting combinations to users based on various factors. 

At its core, this type of model aims to:

- Analyze historical betting data and outcomes
- Consider factors like team/player performance statistics, odds movements, and user preferences
- Make personalized recommendations for potential bets
- Help users construct betting combinations

## Problem Understanding:

The datasets captures user interactions, specifically focusing on betting behavior. Core actor or entities involved:

- Users: The people placing bets
- Bets: Individual wagers placed by users
- Selections: Parts of a bet
- Events: The actual sports matches/games
- Markets: Different types of bets available for each event

Key Relationship:

- One bet can have multiple selections
- Each selection is tied to: A specific event or A specific market
- The same event can appear in many different betslips
- They might even bet on different markets within the same event

Lets break down each table and explain the likely purpose of each column:


### Users Table
list of registered users
- **player_id**: Unique identifier for each player/bettor
- brand_id: Identifier for the betting platform/company (useful if multiple brands exist)
- player_reg_date: Registration date of the player
- language: Preferred language of the player

### Events Table
List of sporting events available for betting
- event_id: Unique identifier for each sporting event
- start_time: When the event/match starts
- sport_id: Identifier for the type of sport (e.g., 1=football, 2=basketball)
- league_id: Identifier for the specific league/competition
- home_team: Name/ID of the home team
- away_team: Name/ID of the away team

### Bets Table
Transactional table providing the hisotrical data on bets. There is relationship between Users and Events
- bet_id: Unique identifier for each bet
- brand_id: Links to the brand where bet was placed
- player_id: Links to the user who placed the bet
- bet_date: When the bet was placed
- amount: Stake amount
- amount_currency: Currency of the bet
- bet_type: Type of bet (e.g., single, accumulator/parlay)
- bet_odds: Total odds for the bet
- status: Current status of bet (e.g., pending, won, lost)
- settlement_timestamp: When the bet was settled/resulted
- outcome_id: Identifier for the specific outcome chosen
- specifier: Additional bet specifications/parameters
- event_id: Links to the specific event
- market_id: Type of market (e.g., match winner, over/under)
- outcome_odds: Odds for this specific selection

import pandas as pd
from recsys.config import Settings
from recsys.utils.utils import IDConverter
from recsys.features import users, events, interactions, labels, BettingDataset
import pickle

settings = Settings()
SOURCE_DIR = settings.SOURCE_DATA_DIR
processed_dir = settings.PROCESSED_DATA_DIR

# Prepare data
user_df = pd.read_csv(SOURCE_DIR / 'users.csv')
event_df = pd.read_csv(SOURCE_DIR / 'events.csv')
bet_df = pd.read_csv(SOURCE_DIR / 'bets.csv')

# Check for nulls
user_df.isnull().sum()


# Check for Duplicates
user_df[user_df.duplicated()]

user_df_copy = user_df.copy()

user_df_copy['date_len'] = user_df_copy.player_reg_date.str.len()

user_df_copy.date_len.value_counts()

user_df_copy['player_reg_date'] = pd.to_datetime(user_df_copy.player_reg_date.str[:19])

user_df = user_df.drop_duplicates().copy(deep=True)
event_df = event_df.drop_duplicates().copy(deep=True)
interactions_df = bet_df.drop_duplicates().copy(deep= True)

user_df.columns

event_df.columns

interactions_df.columns

interaction_cols = ['bet_id', 'brand_id', 'player_id', 'bet_date', 'amount', 'amount_currency', 'bet_type', 'bet_odds', 'status','event_id','market_id', 'outcome_odds']

interactions_df = interactions_df[interaction_cols].merge(user_df[['player_id', 'brand_id',]], how = 'left', left_on='player_id', right_on = 'player_id')

interactions_df.bet_date = pd.to_datetime(interactions_df.bet_date.str[:19])

interactions_df = interactions_df.merge(event_df[['event_id', 'start_time']], how='left', left_on='event_id', right_on='event_id')

interactions_df.sort_values(['start_time','bet_date', 'player_id'], ascending=[False, False, False])

interactions_df.groupby(['player_id', pd.Grouper(key='bet_date', freq='M')]).agg({'bet_id':'count'}).unstack()

interactions_df.pivot_table(columns='bet_type', index='market_id', values= 'bet_id', aggfunc='count')
