==> notebooks/01_compute_features.ipynb <==
# Feature pipeline: Computing the Features

import pandas as pd
from recsys.config import Settings
from recsys.utils.utils import IDConverter
from recsys.features import users, events, interactions, labels, BettingDataset
import pickle

settings = Settings()
SOURCE_DIR = settings.SOURCE_DATA_DIR
processed_dir = settings.PROCESSED_DATA_DIR

# Prepare data
user_df = pd.read_csv(SOURCE_DIR / 'users.csv')
event_df = pd.read_csv(SOURCE_DIR / 'events.csv')
interactions_df = pd.read_csv(SOURCE_DIR / 'bets.csv')

bet_events = set(interactions_df.event_id.to_list())
evnts = set(event_df.event_id.to_list())
list_of_events = evnts.union(bet_events)
event_id_converter = IDConverter()
for id in list_of_events:
    event_id_converter.convert(id)

### IDConverter class
The IDConverter class maintains bidirectional mappings between hex IDs and integers, allowing us to:

- Convert input IDs to integers for model training
- Convert back to original IDs when making recommendations
- Maintain data consistency throughout the pipeline

player_id_converter = IDConverter()
user_df['player_id'] = user_df['player_id'].apply(player_id_converter.convert)
brand_id_converter = IDConverter()
user_df['brand_id'] = user_df['brand_id'].apply(brand_id_converter.convert)
# bet_id, outcome_id, market_id
bet_id_converter = IDConverter()
interactions_df['bet_id'] = interactions_df['bet_id'].apply(bet_id_converter.convert)
outcome_id_converter = IDConverter()
interactions_df['outcome_id'] = interactions_df['outcome_id'].apply(outcome_id_converter.convert)
market_id_converter = IDConverter()
interactions_df['market_id'] = interactions_df['market_id'].apply(market_id_converter.convert)

event_df['event_id'] = event_df['event_id'].map(event_id_converter.id_to_int)
sport_id_converter = IDConverter()
event_df['sport_id'] = event_df['sport_id'].apply(sport_id_converter.convert)
league_id_converter = IDConverter()
event_df['league_id'] = event_df['league_id'].apply(league_id_converter.convert)

interactions_df.player_id = interactions_df.player_id.map(player_id_converter.id_to_int)
interactions_df.brand_id = interactions_df.brand_id.map(brand_id_converter.id_to_int)
interactions_df.event_id = interactions_df.event_id.map(event_id_converter.id_to_int)

user_df = user_df.drop_duplicates().copy(deep=True)
event_df = event_df.drop_duplicates().copy(deep=True)
interactions_df = interactions_df.drop_duplicates().copy(deep= True)
user_df.player_reg_date = user_df.player_reg_date.str[:19]

with open(processed_dir / 'user_df.pickle', 'wb') as handle:
    pickle.dump(user_df, handle, protocol=pickle.HIGHEST_PROTOCOL)

user_df.head()


with open(processed_dir / 'events_df.pickle', 'wb') as handle:
    pickle.dump(event_df, handle, protocol=pickle.HIGHEST_PROTOCOL)

event_df.head()

interactions_df.head()
with open(processed_dir / 'bets_df.pickle', 'wb') as handle:
    pickle.dump(interactions_df, handle, protocol=pickle.HIGHEST_PROTOCOL)

user_features, user_array = users.process_user_features(users_df=user_df, bets_df=interactions_df)

user_features

event_features, event_array = events.process_event_features(events_df=event_df)

event_features

valid_user_ids = users.get_valid_users(users_df=user_df)
valid_event_ids = events.get_valid_events(events_df=event_df)

map_interactions = interactions.process_interactions(bets_df=interactions_df, valid_user_ids=valid_user_ids, valid_event_ids=valid_event_ids)

map_interactions

- Added proper interaction matrix creation
- Created explicit user-event pairs for training

labelled_data = labels.create_labels(interactions=map_interactions, valid_user_ids=valid_user_ids, valid_event_ids=valid_event_ids)

user_array.shape[1]

dataset = BettingDataset(user_df=user_df, event_df=event_df, bets_df=interactions_df)


with open(processed_dir / 'features_dataset.pickle', 'wb') as handle:
    pickle.dump(dataset, handle, protocol=pickle.HIGHEST_PROTOCOL)
