==> notebooks/02_training_2T_model.ipynb <==
# Training pipeline: Training Two Tower model

In this notebook, We will train a retrieval model that will be able to quickly generate a small subset of Users events from a large collection of bets. The Model is based on the two-tower architecture, which embeds Users Inputs (Query) and bets (bet_ids) into a shared low-dimensional vector space. Here, a query consists of features of a Users and a bets (e.g. timestamp of the events), whereas a events consists of features of a particular bet. All queries will have a user ID and all bets will have an bet ID, and the model will be trained such that the embedding of a user will be close to all the embeddings of bets the user has previously bought.

# Imports
import pickle
import torch
from recsys.config import settings
from recsys.train import train_test_split, experiment
from recsys.models import two_tower

# Getting the training data
feature_dir = settings.PROCESSED_DATA_DIR
with open(feature_dir / 'features_dataset.pickle', 'rb') as handle:
    features_data = pickle.load(handle)


train_data, val_data = train_test_split.get_train_val_split(features_data)

model = two_tower.initialize_model(dataset=features_data, device=settings.DEVICE)

criterion, optimizer = two_tower.setup_training(model)

experiment.train_model(model, train_data, val_data, criterion, optimizer, num_epochs= 10)

model.state_dict()

import pathlib

MODEL_REGISTRY = pathlib.Path().cwd().parent / 'model_registry'

model_save_path = MODEL_REGISTRY / 'two_tower_10_001_80_250130.pth'

# Save the model state_dict()
print(f"[INFO] Saving model to: {model_save_path}")
torch.save(obj=model,
            f=model_save_path)
